requirements.txt
langchain==0.2.14
openai
crewai==0.51.1
crewai-tools==0.8.3
python-dotenv